Usability improvements
The tool now has the option to automatically save graphs - though due to how the library works, markers in the graphs (e.g. points in the scatter graphs) are larger when automatically saved.

Along with that, it also has the option to not show the graphs as they are generated, if you just want to generate the graphs as images.

I've also fixed the score histogram to not include battles without results, as that was inflating the left-most bar in the histogram above where it should be.